requirements.txt
psycopg
python-dotenv
requests